+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [0.5.0] - 2026-08-10
+
 ### Changed
 
 - Breaking: `bela-sys` enum constants now use the names from Bela's C
@@ -706,7 +708,8 @@ and this project adheres to
 - Dual MIT / Apache-2.0 licensing
 - A draft of the cross-compilation setup in `docs/cross-compile.md`
 
-[Unreleased]: https://github.com/akiomik/bela-rs/compare/v0.4.0...HEAD
+[Unreleased]: https://github.com/akiomik/bela-rs/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/akiomik/bela-rs/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/akiomik/bela-rs/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/akiomik/bela-rs/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/akiomik/bela-rs/compare/v0.1.0...v0.2.0